Fix bug with dependency on showLineNumber option
This plugin has a dependency on the "Show line number" option in
Obsidian settings.

Before this change, the plugin checked the status of this option during
`onload()`. If the option was OFF the plugin did nothing. If the option
was ON the plugin overwrites the `lineNumberFormatter` function with one
that displays relative line numbers.

`onload()` only runs when turning the plugin on in the settings, or on
Obsidian startup if the plugin has been turned on in a previous setting.

Because the status of the "Show line number" option is only checked at
these times, it is possible for the plugin to get into a broken state:
- "Show line number" is OFF
- User installs this plugin
- User turns the plugin ON
- User turns "Show line number" ON
Until Obsidian is restarted, the plugin will not be working after these
steps.

To fix this, I have added a listener for config changes which inspects
the status of the plugin and the setting, and overwrites or resets the
`lineNumberFormatter` appropriately.

An alternative fix could be to always overwrite `lineNumberFormatter`
and not inspect the setting. I've verified that this doesn't cause any
errors, but it does mean unnecessary work is being done.

This approach to listening to config changes wasn't documented anywhere
that I could find, so there maybe a better way to achieve the same.